+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [8.64.0](https://github.com/doublewordai/control-layer/compare/v8.63.0...v8.64.0) (2026-06-26)
+
+
+### Features
+
+* erase a deleted user's API keys and fusillade data ([#1174](https://github.com/doublewordai/control-layer/issues/1174)) ([c5182c8](https://github.com/doublewordai/control-layer/commit/c5182c8952036823bc0a734dc478382a7860c01c))
+
+## [8.63.0](https://github.com/doublewordai/control-layer/compare/v8.62.0...v8.63.0) (2026-06-25)
+
+
+### Features
+
+* input-caching ([#1172](https://github.com/doublewordai/control-layer/issues/1172)) ([9bf7a0b](https://github.com/doublewordai/control-layer/commit/9bf7a0b5631a3606a6d951ff8600b7a0577be40a))
+
 ## [8.62.0](https://github.com/doublewordai/control-layer/compare/v8.61.3...v8.62.0) (2026-06-25)